Add support for the Lifecycle test
This test is a stability/burn-in test. It is not a speed test. A
receiver federate will sit in a federation and wait for a sender to
continually join/resign. Object creation/update will occur as part of
each join/resign cycle, ensuring that there is communication between the
federates.

This test ensures that the wiring up of all comms infrastructure works
over time and over many join/resign cycles.